Temporary fix for incorrect data: uppercase/lowercase image file extensions
Many images (thousands) have an uppercase file extension in the DB,
when the file has a lowercase extension. Since image manipulation
was made on a case-insentive file system, we have inconsistent cases
depending on the file. At 3 days from launch. With over 300GB of data 🤯.
This will be solved when images are moved to a proper digital asset
management system (DAMS), currently underway.