Reimplement the selections grid, to work within constraints of masonry.
The masonry library (Bricks.js) requires, for performance, that
the gutters are fixed, and that the widths of grid items are
explicitly set in CSS.
Because the design was using fluid columns as gutters (and gutters
as columns), this requirement wasn't being met.
This forced me to reimplement the integration of the selections grid,
without using fluid gutters, while respecting as much as possible the
spirit (if not the letter) of the visual design.